Relocated some purely numpy-based functions to numerics.
