+++ src/dymad/numerics/__init__.py
@@ -1,0 +1,7 @@
+from dymad.numerics.linalg import truncated_svd
+from dymad.numerics.weak import generate_weak_weights
+
+__all__ = [
+    "generate_weak_weights",
+    "truncated_svd",
+]
