+++ src/dymad/numerics/weak.py
