+++ src/dymad/utils/__init__.py
@@ -1,11 +1,9 @@
 from dymad.utils.checkpoint import load_checkpoint, load_model, save_checkpoint
-from dymad.utils.linalg import truncated_svd
 from dymad.utils.misc import load_config, setup_logging
 from dymad.utils.modules import ControlInterpolator, FlexLinear, GNN, IdenCatGNN, IdenCatMLP, make_autoencoder, MLP, ResBlockGNN, ResBlockMLP
 from dymad.utils.plot import plot_summary, plot_trajectory, plot_hist
 from dymad.utils.prediction import predict_continuous, predict_discrete, predict_graph_continuous, predict_graph_discrete
 from dymad.utils.preprocessing import Compose, DelayEmbedder, Identity, make_transform, Scaler, SVD
-from dymad.utils.weak import generate_weak_weights
 from dymad.utils.sampling import TrajectorySampler
 from dymad.utils.scheduler import make_scheduler
 
@@ -14,7 +12,6 @@ __all__ = [
     "ControlInterpolator",
     "DelayEmbedder",
     "FlexLinear",
-    "generate_weak_weights",
     "GNN",
     "IdenCatGNN",
     "IdenCatMLP",
@@ -40,5 +37,4 @@ __all__ = [
     "setup_logging",
     "SVD",
     "TrajectorySampler",
-    "truncated_svd",
 ]
